+++ jest.config copy.js
@@ -1,0 +1,10 @@
+module.exports = {
+  preset: 'ts-jest',
+  testEnvironment: 'jsdom',
+  maxWorkers: 4,
+  globals: {
+    'ts-jest': {
+      isolatedModules: true,
+    },
+  },
+}